requirements.txt
confluent-kafka>=2.4.0
python-dotenv>=1.0.1
google-cloud-bigquery>=3.25.0
